Split `CMOS` and `RTC` into different modules.
